requirements.txt
mysql-connector-python
colorama~=0.4.4
requests~=2.25.1
manuf
pycryptodome